Create reusable `InputStartDate` component
